implemented PaymentCalculatorService

+++ ParkingManagement.Api/ParkingManagement.Domain/Dtos/PaymentDto.cs
@@ -3,7 +3,7 @@
     public class PaymentDto
     {
         public TimeSpan Duration { get; set; }
-        public int NumberOfHoursToPay { get; set; }
+        public double NumberOfHoursToPay { get; set; }
         public decimal PriceBaseValue { get; set; }
         public decimal TotalPayable { get; set; }
     }
